- use memmove, not memcpy.  avoids warnings. - vram control registers are 8bit, not 32bit.  setting VramReg.acr would clear bcr,ccr,dcr...  also remove a few unused variables. - zero-terminate the buffer that holds the user name and message (from the firmware).  change the type of these buffers to Rune.

+++ io.h
@@ -76,15 +76,15 @@ struct LcdReg {
 #define VRAMREG	((VramReg*)VRAM)
 typedef struct VramReg VramReg;
 struct VramReg {
-	ulong acr;
-	ulong bcr;
-	ulong ccr;
-	ulong dcr;
-	ulong ecr;
-	ulong fcr;
-	ulong gcr;
-	ulong hcr;
-	ulong icr;
+	uchar acr;
+	uchar bcr;
+	uchar ccr;
+	uchar dcr;
+	uchar ecr;
+	uchar fcr;
+	uchar gcr;
+	uchar hcr;
+	uchar icr;
 };
 
 #define POWERREG ((PowerReg*)POWER)
